JGSS: Do not reacquire cred upon context establishment
We were reacquiring the initiator/acceptor credential upon security
context full establishment.  This is pointless.  It doesn't achieve
anything other than a) consume more memory with iceberg objects (the JVM
doesn't know how large they are), b) it breaks with C native GSS
implementations because the reacquisition is notnecessarily done with
the same mechanism.

(b) happens with Heimdal in particular, where we accept a context with
some mechanism (e.g., SPNEGO), get a different actual mech, and then we
try and fail to acquire a credential for the accepted targetName with
the wrong mechanism, which then raise a GSSEception.